feature - Final part of the card design
As the research goes on, we realized that image services such as
imgur, sidebar.io is not using Masonry. This is because Masonry like
UI is hard for user to use in such service. We decided that using
Masonry on this app is not the right approach and tried to use the
card ui on Materialize.
imports/style.css
imports/ui/partials/Memo.html