g36 - Add options for memos
Description
We removed updating label method from memos view long time ago. We want
them back. We also want to favorite memos for future feature.

Solution
Added card-action section which has updating memo and favorite button.
Because of that, we adjusted the styling for the memo.html.
We should add clickCount method on this branch beacuase it is easy to
implement.