g149 - Fix requestUser error
Description
Fix requestUser error occuring in the production by doing the user
validation only on server.
The problem was that Meteor users can be queried only on server side
and the publishment was not being reactive to the labelShare.

Solution
Fix the meteor method so that query will occur only on server side.
Use the publishComposite so that user collections will be reactive
to the labelShare.